fix: add scrollMode and currentPage fixes when change between modes
